Move shared Node.js payload logic to mixin. - this fixes the recursive loading issue when creating a payload   inside the cmd payload - also dries up some of the node cmd invocation logic.
